[ghstack-poisoned]

+++ tests/python/test_argument.py
@@ -224,7 +224,7 @@ def test_args_with_many_ndarrays():
 
 @test_utils.test()
 def test_struct_arg():
-    s0 = ti.types.struct(a=ti.i16, b=ti.f32)
+    s0 = ti.types.struct(a=ti.i16, b=ti.f64)
     s1 = ti.types.struct(a=ti.f32, b=s0)
 
     @ti.kernel
